refactor: refactored structure of checkers into multiple files
